clean ieee measurement blocks

+++ example/ieee802_11/gnuradio_blocks/gr-gnugym/lib/equalizer/comb.h
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2015 Bastian Bloessl <bloessl@ccs-labs.org>
+ * and 2020 Sascha Rösler, TU Berlin <s.roesler@campus.tu-berlin.de>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -26,7 +27,7 @@ namespace equalizer {
 
 class comb: public base {
 public:
-	virtual void equalize(gr_complex *in, int n, gr_complex *symbols, uint8_t *bits, boost::shared_ptr<gr::digital::constellation> mod, double* snrVect, bool onlyRssi);
+	virtual void equalize(gr_complex *in, int n, gr_complex *symbols, uint8_t *bits, boost::shared_ptr<gr::digital::constellation> mod, float* snrVect, bool onlyRssi);
 	double get_snr();
 
 private:
